♻️  zu: Move Signature to the top `signature` mod
This makes sense anyway but it also avoid a clippy warning:

```rust
error: module has the same name as its containing module
 --> zvariant_utils/src/signature/mod.rs:8:1
  |
8 | pub mod signature;
  | ^^^^^^^^^^^^^^^^^^
  |
  = help: for further information visit https://rust-lang.github.io/rust-clippy/master/index.html#module_inception
  = note: `-D clippy::module-inception` implied by `-D warnings`
  = help: to override `-D warnings` add `#[allow(clippy::module_inception)]`
```
