+++ zvariant_utils/src/signature/tests.rs
@@ -1,4 +1,4 @@
-use super::{signature::*, *};
+use super::*;
 use std::{
     hash::{DefaultHasher, Hash, Hasher},
     str::FromStr,
